split quiz into multiple files
