+++ src/pages/quiz/quiz.tsx
@@ -2,26 +2,24 @@ import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import useLocalStorageState from "use-local-storage-state";
 import useSessionStorageState from "use-session-storage-state";
-import { Responsive as ResponsiveGridLayout } from "react-grid-layout";
 import { FaCog, FaHome, FaPencilAlt, FaPlay, FaRedo } from "react-icons/fa";
 import { BsFullscreen, BsFullscreenExit } from "react-icons/bs";
 import { quizLayout } from "./quizLayout";
 import { motion } from "framer-motion";
-import { Layout, QuizBackend, Streamer, TriviaSettings } from "../../types";
-import ChatComponent from "../../components/chatComponent/chatComponent";
-import QuizComponent from "../../components/quizComponent/quizComponent";
-import AnnouncementFeedComponent from "../../components/announcementFeedComponent/announcementFeedComponent";
+import { QuizBackend, TriviaSettings } from "../../types";
 import "./quiz.css";
 import toast from "react-hot-toast";
 import ResolvableToast from "../../components/toasts/resolvableToast";
-import clsx from "clsx";
-import { MdDragIndicator } from "react-icons/md";
 import { io, Socket } from "socket.io-client";
 import { appConfig } from "../../config/app";
+import QuizLogic from "./quizLogic";
+import QuizGrid from "./quizGrid";
+import QuitQuizToast from "./quitQuizToast";
 
 export default function Quiz() {
   const navigate = useNavigate();
 
+  //retrieve account from session storage
   const [account] = useSessionStorageState("account", {
     defaultValue: {
       username: "",
@@ -29,16 +27,12 @@ export default function Quiz() {
       uid: "",
     },
   });
-  const streamer: Streamer = {
-    channel: account.username,
-    uid: account.uid,
-  };
 
   //initial settings
   const [triviaSettings, setTriviaSettings] =
     useLocalStorageState<TriviaSettings>("quizSettings", {
       defaultValue: {
-        channel: streamer.channel,
+        channel: account.username,
         startAfter: 0.000000000001,
         questionAmount: 10,
         timePerQuestion: 15,
@@ -49,7 +43,7 @@ export default function Quiz() {
       },
     });
 
-  //set channel to streamer channel (temporary)
+  //set channel to streamer channel (temporary, want to change to channel id?)
   setTriviaSettings((prevState) => ({
     ...prevState,
     channel: account.username,
@@ -57,8 +51,6 @@ export default function Quiz() {
 
   const [start, setStart] = useState(false);
   const [connect, setConnect] = useState(false);
-
-  //show toast when user tries to leave page, only if quiz has started. use window.history to prevent user from going back before he has answered the toast
 
   //get quiz data from back-end
   const [quiz, setQuiz] = useState<QuizBackend>({
@@ -87,86 +79,31 @@ export default function Quiz() {
     setSocket(socket);
   }, []);
 
-  function disConnect() {
+  function disconnect() {
     console.log("disconnecting");
     console.log(socket);
     socket?.disconnect();
     console.log(socket);
   }
 
-  useEffect(() => {
-    if (start) {
-      if (!socket) return;
-      //on first connection, send quiz to back-end
-      socket.emit("trivia-start", triviaSettings);
-
-      //give countdown before first question
-      socket.on("game-starting", (data) => {
-        setQuiz((prevState) => ({
-          ...prevState,
-          time: data.in,
-        }));
-        //confirm the connection with back-end
-        setConnect(true);
-      });
-
-      //when getting a new question, update the data
-      socket.on("question-new", (data) => {
-        console.log(data);
-        setQuiz((prevState) => ({
-          ...prevState,
-          question: data.question,
-          possibilities: data.possibilities,
-          rightAnswer: "",
-          percentages: [],
-          questionIndex: data.questionIndex,
-          time: data.time,
-        }));
-      });
-
-      //after {timePerQuestion} show the right answer and the percentages
-      socket.on("question-finished", (data) => {
-        console.log(data);
-        setQuiz((prevState) => ({
-          ...prevState,
-          rightAnswer: data.rightAnswer,
-          percentages: data.percentages,
-          announcements: {
-            mostPoints: data.mostPoints?.username ?? "",
-            mostPointsAmount: data.mostPoints?.points ?? "",
-            firstToGuess: data.firstToGuess,
-            onStreak: data.contestantData[0]?.username ?? "",
-            onStreakAmount: data.contestantData[0]?.currentStreak ?? "",
-          },
-        }));
-      });
-
-      //when the game is finished, show the results
-      socket.on("game-finished", (data) => {
-        setQuiz((prevState) => ({
-          ...prevState,
-          results: data.results,
-        }));
-        console.log(data.results);
-        //wait 5 seconds before navigating to the results page
-        setTimeout(() => {
-          toast.dismiss();
-          disConnect();
-          navigate("/quizresults", {
-            state: {
-              results: data.results,
-            },
-          });
-        }, 5000);
-      });
-    }
-  }, [start]);
+  //the quizLogic function is a custom function that handles all the socket logic
+  QuizLogic({
+    socket,
+    start,
+    setConnect,
+    setQuiz,
+    triviaSettings,
+    disconnect,
+  });
 
   window.onpopstate = function () {
-    disConnect();
+    disconnect();
   };
 
-  //useLocalStorageState hook to save the layout
+  const [editable, setEditable] = useState(false);
+  const [fullscreen, setFullscreen] = useState(false);
+
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const [layout, setLayout, { removeItem }] = useLocalStorageState(
     "quizLayout",
     {
@@ -174,20 +111,6 @@ export default function Quiz() {
     }
   );
 
-  const [editable, setEditable] = useState(false);
-  const [fullscreen, setFullscreen] = useState(false);
-
-  // update height every time the window.innerHeight changes
-  const [height, setHeight] = useState(window.innerHeight - 20);
-  useEffect(() => {
-    const handleResize = () => {
-      setHeight(window.innerHeight - 20);
-    };
-    window.addEventListener("resize", handleResize);
-    return () => {
-      window.removeEventListener("resize", handleResize);
-    };
-  }, []);
   return (
     <motion.div
       initial={{
@@ -209,7 +132,19 @@ export default function Quiz() {
         {/* create menu button */}
         <button
           onClick={() => {
-            toast.success("Game started");
+            toast.promise(
+              new Promise<void>((resolve) => {
+                //wait for socket.game-starting
+                socket?.on("game-starting", () => {
+                  resolve();
+                });
+              }),
+              {
+                loading: "Starting game",
+                success: "Game started",
+                error: "Something went wrong",
+              }
+            );
             setStart(true);
           }}
           className="group relative flex h-[75px] w-[75px] items-center justify-center overflow-hidden whitespace-nowrap rounded-babble border border-babbleGray bg-babbleLightGray/5 p-4 text-white shadow-babbleOuter backdrop-blur-babble hover:overflow-hidden hover:border-babbleOrange hover:text-babbleWhite"
@@ -220,26 +155,7 @@ export default function Quiz() {
         <button
           onClick={() => {
             if (start) {
-              toast.loading(
-                (t) => (
-                  <ResolvableToast
-                    t={t}
-                    text="Are you sure you want to quit?"
-                    confirm="Quit"
-                    cancel="Continue"
-                    func={() => {
-                      setStart(false);
-                      disConnect();
-                      navigate("/settings");
-                      //add this page to the history so the user can go back to it
-                    }}
-                  />
-                ),
-                {
-                  icon: <></>,
-                  id: "exit",
-                }
-              );
+              QuitQuizToast({ setStart, disconnect });
             } else {
               navigate("/settings");
             }
@@ -252,27 +168,9 @@ export default function Quiz() {
         <button
           onClick={() => {
             if (start) {
-              toast.loading(
-                (t) => (
-                  <ResolvableToast
-                    t={t}
-                    text="Are you sure you want to quit?"
-                    confirm="Quit"
-                    cancel="Continue"
-                    func={() => {
-                      setStart(false);
-                      disConnect();
-                      navigate("/");
-                    }}
-                  />
-                ),
-                {
-                  icon: <></>,
-                  id: "exit",
-                }
-              );
+              QuitQuizToast({ setStart, disconnect });
             } else {
-              disConnect();
+              disconnect();
               console.log("disconnected");
               navigate("/");
             }
@@ -283,73 +181,16 @@ export default function Quiz() {
           <div className="absolute inset-0 z-0 h-full w-full overflow-hidden bg-gradient-to-br from-babbleOrange/20 to-babbleOrange/0 opacity-0 transition duration-300 hover:opacity-100 group-hover:opacity-100" />
         </button>
       </div>
-      <ResponsiveGridLayout
-        className="overflow-hidden"
-        width={window.innerWidth}
-        layouts={layout}
-        breakpoints={{ lg: 100 }}
-        cols={{ lg: 24 }}
-        rowHeight={height / 12 - 52.75}
-        isBounded={true}
-        useCSSTransforms={true}
-        // allowOverlap={true}
-        compactType={"vertical"}
-        resizeHandles={["se"]}
-        margin={[50, 50]}
-        isResizable={editable && !start}
-        isDraggable={editable && !start}
-        onLayoutChange={(layout: Layout[]) => {
-          setLayout((prevState) => ({
-            ...prevState,
-            lg: layout,
-          }));
-        }}
-      >
-        <div
-          className={clsx(
-            "relative z-10 flex w-[450px] items-center justify-center ",
-            editable && !start && "cursor-grab"
-          )}
-          key="chat-component"
-        >
-          {editable && !start && (
-            <MdDragIndicator className="absolute inset-y-2 left-2 z-20 m-auto text-xl text-white" />
-          )}
-          <ChatComponent
-            streamer={streamer}
-            platform={account.platform}
-            announcements={quiz.announcements}
-            socket={socket}
-          />
-        </div>
-        <div
-          className={clsx(
-            "relative z-10 flex items-center justify-center",
-            editable && !start && "cursor-grab"
-          )}
-          key="quiz-component"
-        >
-          {editable && !start && (
-            <MdDragIndicator className="absolute inset-y-2 left-2 z-20 m-auto text-xl text-white" />
-          )}
-          <QuizComponent quiz={quiz} start={connect} />
-        </div>
-        <div
-          className={clsx(
-            "relative z-10 flex items-center justify-center",
-            editable && !start && "cursor-grab"
-          )}
-          key="first-to-answer"
-        >
-          {editable && !start && (
-            <MdDragIndicator className="absolute inset-y-2 left-2 z-20 m-auto text-xl text-white" />
-          )}
-          <AnnouncementFeedComponent
-            key="first-to-answer"
-            announcements={quiz.announcements}
-          />
-        </div>
-      </ResponsiveGridLayout>
+      <QuizGrid
+        editable={editable}
+        socket={socket}
+        account={account}
+        quiz={quiz}
+        start={start}
+        streamer={account}
+        connect={connect}
+      />
+
       <div className="absolute bottom-[50px] left-[50px] z-40 flex flex-col gap-[25px] text-[25px] font-[1000]">
         {editable && !start && (
           <button
